+++ src/pluginsettings.hpp
@@ -15,6 +15,7 @@ struct Settings {
 	bool mbFavoriteHighlight = true;
 	float mbSearchThreshold = 0.5f;
 	bool mbMagnifierEnabled = false;
+	bool mbArrowKeyNavigation = true;
 	bool mbPrewarmEnabled = false;
 	bool mbApplyLibraryWhitelist = false;
 	bool mbShowDeprecated = false;