+++ CHANGELOG.md
@@ -22,6 +22,7 @@
     - Improved rendering performance
 - Module [MB](./docs/trial-and-error/Mb.md)
     - Added option to import "recently used"/"most used" statistics from Rack's own module browser
+    - Added option to disable arrow keys selecting modules in *v2 mod* browser
     - Added option to pre-warm module previews
     - Fixed "Newest" module manifest download if Rack user folder is on different disk
 - Module [SIREN](./docs/siren/Siren.md)